# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ComplianceAI API
  description: AI-powered cybersecurity documentation automation platform
  version: 1.0.0
  contact:
    name: ComplianceAI Support
    email: support@complianceai.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://production-url.replit.app/api
    description: Production server
security:
  - sessionAuth: []
paths:
...
  /auth/user:
    get:
      summary: Get current user
      description: Returns the authenticated user's profile
      tags:
        - Authentication
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /organizations:
    get:
      summary: Get user organizations
      description: Returns organizations the user is a member of
      tags:
        - Organizations
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
    post:
      summary: Create organization
      description: Creates a new organization
      tags:
        - Organizations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganization'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
  /company-profiles:
    get:
      summary: Get company profiles
      description: Returns a list of company profiles
      tags:
        - Company Profiles
      parameters:
        - name: organizationId
          in: query
          description: Filter by organization ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of company profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyProfile'
    post:
      summary: Create company profile
      description: Creates a new company profile
      tags:
        - Company Profiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyProfile'
      responses:
        '201':
          description: Company profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfile'
  /company-profiles/{id}:
    get:
      summary: Get company profile
      description: Returns a specific company profile
      tags:
        - Company Profiles
      parameters:
        - name: id
          in: path
          required: true
          description: Company profile ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Company profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfile'
        '404':
          description: Company profile not found
    put:
      summary: Update company profile
      description: Updates a company profile
      tags:
        - Company Profiles
      parameters:
        - name: id
          in: path
          required: true
          description: Company profile ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyProfile'
      responses:
        '200':
          description: Company profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfile'
  /documents:
    get:
      summary: Get documents
      description: Returns a list of documents
      tags:
        - Documents
      parameters:
        - name: companyProfileId
          in: query
          description: Filter by company profile ID
          schema:
            type: string
            format: uuid
        - name: framework
          in: query
          description: Filter by framework
          schema:
            type: string
            enum: [ISO27001, SOC2, FedRAMP, NIST]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        profileImageUrl:
          type: string
          format: uri
        role:
          type: string
          enum: [user, admin, org_admin]
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        logo:
          type: string
          format: uri
        website:
          type: string
          format: uri
        contactEmail:
          type: string
          format: email
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrganization:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 1
          maxLength: 100
        description:
          type: string
        website:
          type: string
          format: uri
        contactEmail:
          type: string
          format: email
    CompanyProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        companyName:
          type: string
        industry:
          type: string
        companySize:
          type: string
        headquarters:
          type: string
        cloudInfrastructure:
          type: array
          items:
            type: string
        dataClassification:
          type: string
        businessApplications:
          type: string
        complianceFrameworks:
          type: array
          items:
            type: string
        contactInfo:
          type: object
        keyPersonnel:
          type: object
        frameworkConfigs:
          type: object
        uploadedDocs:
          type: object
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCompanyProfile:
      type: object
      required:
        - organizationId
        - createdBy
        - companyName
        - industry
        - companySize
        - headquarters
        - dataClassification
        - businessApplications
      properties:
        organizationId:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        companyName:
          type: string
          minLength: 1
          maxLength: 255
        industry:
          type: string
        companySize:
          type: string
        headquarters:
          type: string
        cloudInfrastructure:
          type: array
          items:
            type: string
        dataClassification:
          type: string
        businessApplications:
          type: string
        complianceFrameworks:
          type: array
          items:
            type: string
    UpdateCompanyProfile:
      type: object
      properties:
        companyName:
          type: string
          minLength: 1
          maxLength: 255
        industry:
          type: string
        companySize:
          type: string
        headquarters:
          type: string
        cloudInfrastructure:
          type: array
          items:
            type: string
        dataClassification:
          type: string
        businessApplications:
          type: string
        complianceFrameworks:
          type: array
          items:
            type: string
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyProfileId:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        framework:
          type: string
        category:
          type: string
        documentType:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, in_progress, complete, approved, published]
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentList:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer